pyproject.toml
[project]
name = "agent-pod-lab"
version = "0.1.0"
description = "LLM-orchestrated ML experiment system with DAG scheduling on RunPod GPU pods"
requires-python = ">=3.11"
dependencies = [
    "matrixpfn>=0.1.13",
    "pyyaml>=6.0",
    "runpod>=1.7.0",
    "requests>=2.31.0",
    "boto3>=1.34.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "gunicorn>=23.0.0",
]

[tool.setuptools.packages.find]
include = ["orchestrator*", "api*", "lib*"]
